# perf: Only fetch `/v4/profile/grants` if a user is `restricted`